weather_data_analysis/requirements.txt
asttokens==2.4.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
ipython==8.22.2
jedi==0.19.1
matplotlib-inline==0.1.6
numpy==1.26.4
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
stack-data==0.6.3
traitlets==5.14.1
tzdata==2024.1
wcwidth==0.2.13
